warning fixes & bumped alpha version

+++ util.c
@@ -338,7 +338,7 @@ void empty_buffer(Bit_stream_struc *bs)
 
 int copy_buffer(char *buffer,int size,Bit_stream_struc *bs)
 {
-  size_t minimum = bs->buf_byte_idx + 1;
+  int minimum = bs->buf_byte_idx + 1;
   if (minimum <= 0) return 0;
   if (size!=0 && minimum>size) return -1; /* buffer is too small */
   memcpy(buffer,bs->buf,minimum);
